+++ StatementRepository.php
@@ -23,18 +23,10 @@ use XApi\Repository\Api\Mapping\MappedStatement;
  *
  * @author Christian Flothmann <christian.flothmann@xabbuh.de>
  */
-abstract class StatementRepository
+abstract class StatementRepository implements StatementRepositoryInterface
 {
     /**
-     * Finds a {@link Statement} by id.
-     *
-     * @param string     $statementId The statement id to filter by
-     * @param Actor|null $authority   (Optional) actor that must be the authority
-     *                                of the returned statement
-     *
-     * @return Statement The statement
-     *
-     * @throws NotFoundException if no Statement with the given UUID does exist
+     * {@inheritdoc}
      */
     final public function findStatementById($statementId, Actor $authority = null)
     {
@@ -60,17 +52,7 @@ abstract class StatementRepository
     }
 
     /**
-     * Finds a voided {@link Statement} by id.
-     *
-     * @param string     $voidedStatementId The voided statement id to filter
-     *                                      by
-     * @param Actor|null $authority         (Optional) actor that must be the
-     *                                      authority of the returned statement
-     *
-     * @return Statement The statement
-     *
-     * @throws NotFoundException if no voided Statement with the given UUID
-     *                           does exist
+     * {@inheritdoc}
      */
     final public function findVoidedStatementById($voidedStatementId, Actor $authority = null)
     {
@@ -96,14 +78,7 @@ abstract class StatementRepository
     }
 
     /**
-     * Finds a collection of {@link Statement Statements} filtered by the given
-     * criteria.
-     *
-     * @param StatementsFilter $criteria  The criteria to filter by
-     * @param Actor|null       $authority (Optional) actor that must be the
-     *                                    authority of the returned statements
-     *
-     * @return Statement[] The statements
+     * {@inheritdoc}
      */
     final public function findStatementsBy(StatementsFilter $criteria, Actor $authority = null)
     {
@@ -124,14 +99,7 @@ abstract class StatementRepository
     }
 
     /**
-     * Writes a {@link Statement} to the underlying data storage.
-     *
-     * @param Statement $statement The statement to store
-     * @param bool      $flush     Whether or not to flush the managed objects
-     *                             immediately (i.e. write them to the data
-     *                             storage)
-     *
-     * @return string The UUID of the created Statement
+     * {@inheritdoc}
      */
     final public function storeStatement(Statement $statement, $flush = true)
     {