+++ StatementRepositoryInterface.php
@@ -4,6 +4,7 @@ namespace XApi\Repository\Api;
 
 use Xabbuh\XApi\Model\Actor;
 use Xabbuh\XApi\Model\Statement;
+use Xabbuh\XApi\Model\StatementId;
 use Xabbuh\XApi\Model\StatementsFilter;
 use XApi\Repository\Api\Exception\NotFoundException;
 
@@ -17,30 +18,30 @@ interface StatementRepositoryInterface
     /**
      * Finds a {@link Statement} by id.
      *
-     * @param string     $statementId The statement id to filter by
-     * @param Actor|null $authority   (Optional) actor that must be the authority
-     *                                of the returned statement
+     * @param StatementId $statementId The statement id to filter by
+     * @param Actor|null  $authority   (Optional) actor that must be the authority
+     *                                 of the returned statement
      *
      * @return Statement The statement
      *
      * @throws NotFoundException if no Statement with the given UUID does exist
      */
-    public function findStatementById($statementId, Actor $authority = null);
+    public function findStatementById(StatementId $statementId, Actor $authority = null);
 
     /**
      * Finds a voided {@link Statement} by id.
      *
-     * @param string     $voidedStatementId The voided statement id to filter
-     *                                      by
-     * @param Actor|null $authority         (Optional) actor that must be the
-     *                                      authority of the returned statement
+     * @param StatementId $voidedStatementId The voided statement id to filter
+     *                                       by
+     * @param Actor|null  $authority         (Optional) actor that must be the
+     *                                       authority of the returned statement
      *
      * @return Statement The statement
      *
      * @throws NotFoundException if no voided Statement with the given UUID
      *                           does exist
      */
-    public function findVoidedStatementById($voidedStatementId, Actor $authority = null);
+    public function findVoidedStatementById(StatementId $voidedStatementId, Actor $authority = null);
 
     /**
      * Finds a collection of {@link Statement Statements} filtered by the given
@@ -62,7 +63,7 @@ interface StatementRepositoryInterface
      *                             immediately (i.e. write them to the data
      *                             storage)
      *
-     * @return string The UUID of the created Statement
+     * @return StatementId The id of the created Statement
      */
     public function storeStatement(Statement $statement, $flush = true);
 }