+++ StatementRepository.php
@@ -12,11 +12,11 @@
 namespace XApi\Repository\Api;
 
 use Rhumsaa\Uuid\Uuid;
+use Xabbuh\XApi\Common\Exception\NotFoundException;
 use Xabbuh\XApi\Model\Actor;
 use Xabbuh\XApi\Model\Statement;
 use Xabbuh\XApi\Model\StatementId;
 use Xabbuh\XApi\Model\StatementsFilter;
-use XApi\Repository\Api\Exception\NotFoundException;
 use XApi\Repository\Api\Mapping\MappedStatement;
 
 /**