+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
 * made the package compatible with `3.x` releases of `ramsey/uuid`
 * allow `2.x` releases of the `php-xapi/model` package too
+* Added a `ActivityRepositoryInterface` that defines the public API of a
+  activity repository.
 
 0.3.0
 -----