+++ CHANGELOG.md
@@ -4,8 +4,12 @@ CHANGELOG
 0.4.0
 -----
 
+* dropped suppport for PHP < 5.6 and HHVM
+
 * made the package compatible with `3.x` releases of `ramsey/uuid`
+
 * allow `2.x` releases of the `php-xapi/model` package too
+
 * added an `ActivityRepositoryInterface` that defines the public API of
   an activity repository